requirements.txt
flask==3.0.0
google-cloud-bigquery==3.11.0
